added last hook for refactoring table data. Hook to edit row
